refactor dialogflow_sample
