+++ dialogflow-cx/dialogflow_sample.py
@@ -1,0 +1,28 @@
+class DialogflowSample:
+    """Base class for samples"""
+
+    def set_auth_delegator(self, auth_delegator):
+        self._auth_delegator = auth_delegator
+
+    def set_agent_delegator(self, agent_delegator):
+        self._agent_delegator = agent_delegator
+
+    @property
+    def auth_delegator(self):
+        return self._auth_delegator
+
+    @property
+    def agent_delegator(self):
+        return self._agent_delegator
+
+    @property
+    def project_id(self):
+        return self.auth_delegator.project_id
+
+    @property
+    def location(self):
+        return self.auth_delegator.location
+
+    @property
+    def start_flow(self):
+        return self.agent_delegator.start_flow
